services/pdf-api/requirements.txt
certifi==2023.5.7
kafka-python==2.0.2
minio==7.1.14
pdf2image==1.16.3
Pillow==9.5.0
urllib3==2.0.2
